+++ src/constants/index.ts
@@ -1,1 +1,3 @@
 export { API_ENDPOINTS } from "./api-endpoints";
+export { DEFAULT_ICE_SERVERS } from "./ice-servers";
+